🏷️ channelSchedule 타입 개별 파일화
